+++ src/modules/home/api/list-accommodations.api.ts
@@ -1,7 +1,9 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { getAccommodations } from '../utils/get-accommodations.util';
-import { getNextPageNumber } from '../utils/get-next-page-number.util';
+import { apiClient } from '#/shared/libs/api-client.lib';
+import { getNextPageNumber } from '../utils/pagination.util';
+import { type Accommodation } from '../types/accommodation.type';
 import { type ListAccommodationQueryParams } from '../schemas/list-accommodation-query-params.schema';
+import { type AccommodationListResponseDataMetadata } from '../types/accommodation-list-metadata.type';
 
 export function useListAccommodationsQuery(limit: number, params: ListAccommodationQueryParams) {
   const listAccommodationsQuery = useInfiniteQuery({
@@ -12,4 +14,25 @@ export function useListAccommodationsQuery(limit: number, params: ListAccommodat
   });
 
   return listAccommodationsQuery;
+}
+
+interface AccommodationListResponseData {
+  result: Accommodation[];
+  metadata: AccommodationListResponseDataMetadata;
+}
+
+interface AccommodationListResponse {
+  status: number;
+  message: string;
+  data: AccommodationListResponseData;
+}
+
+export async function getAccommodations(pageParam: number, limit: number, params: ListAccommodationQueryParams) {
+  const searchParams = new URLSearchParams(params).toString();
+
+  const response = await apiClient.get<AccommodationListResponse>(
+    `/accommodations?page=${pageParam}&limit=${limit}&${searchParams}`
+  );
+
+  return response.data.data;
 }
