+++ src/modules/home/components/accommodation-list.tsx
@@ -1,37 +1,33 @@
 import Grid2 from '@mui/material/Grid2';
 import CardImage from '#/assets/images/card-temp-image.jpg';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useRef } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { AccommodationCard } from './accommodation-card';
 import { useListAccommodationsQuery } from '../api/list-accommodations.api';
 import { SkeletonList } from './skeleton-list';
 import { CardSkeleton } from './card-skeleton';
-import { calculateInitialLimit } from '../utils/calculate-initial-limit.helper';
 import { useInView } from 'react-intersection-observer';
 import { useListAccommodationQueryParams } from '#/shared/hooks/list-accommodations-query-params.hook';
+import { useInfiniteScroll } from '#/shared/hooks/infinite-scroll.hook';
+import { usePaginationLimit } from '#/shared/hooks/pagination-limit.hook';
 
 export function AccommodationList() {
-  const [limit, setLimit] = useState(20);
   const { ref: bottomOfPageRef, inView: isBottomOfPageInView } = useInView();
   const { validatedQueryParams } = useListAccommodationQueryParams();
   const cardRef = useRef<HTMLDivElement>(null);
 
+  const { limit } = usePaginationLimit({ cardRef });
   const { fetchNextPage, hasNextPage, isFetchingNextPage, status, data } = useListAccommodationsQuery(
     limit,
     validatedQueryParams
   );
 
-  useEffect(() => {
-    const initialNumberOfCards = calculateInitialLimit(cardRef.current, limit);
-
-    setLimit(initialNumberOfCards);
-  }, [limit]);
-
-  useEffect(() => {
-    if (isBottomOfPageInView && hasNextPage && !isFetchingNextPage) {
-      fetchNextPage();
-    }
-  }, [isBottomOfPageInView, hasNextPage, fetchNextPage, isFetchingNextPage]);
+  useInfiniteScroll({
+    isBottomOfPageInView,
+    hasNextPage,
+    isFetchingNextPage,
+    fetchNextPage,
+  });
 
   if (status === 'error') {
     return <p>{"Couldn't load data"}</p>;
