+++ src/modules/home/types/accommodation-list-metadata.type.ts
@@ -1,0 +1,6 @@
+export interface AccommodationListResponseDataMetadata {
+  page: number;
+  hasNextPage: boolean;
+  hasPreviousPage: boolean;
+  total: number;
+}
