--- src/modules/home/utils/get-next-page-number.util.ts
@@ -1,9 +1,0 @@
-import { type AccommodationListResponseDataMetadata } from './get-accommodations.util';
-
-export function getNextPageNumber(metadata: AccommodationListResponseDataMetadata) {
-  if (!metadata.hasNextPage) {
-    return undefined;
-  }
-
-  return metadata.page + 1;
-}
