+++ src/shared/hooks/infinite-scroll.hook.tsx
@@ -1,0 +1,18 @@
+import { useEffect } from 'react';
+
+interface InfiniteScrollProps {
+  isBottomOfPageInView: boolean;
+  hasNextPage: boolean;
+  isFetchingNextPage: boolean;
+  fetchNextPage: () => void;
+}
+
+const useInfiniteScroll = ({ isBottomOfPageInView, hasNextPage, fetchNextPage, isFetchingNextPage }: InfiniteScrollProps) => {
+  useEffect(() => {
+    if (isBottomOfPageInView && hasNextPage && !isFetchingNextPage) {
+      fetchNextPage();
+    }
+  }, [isBottomOfPageInView, hasNextPage, fetchNextPage, isFetchingNextPage]);
+};
+
+export { useInfiniteScroll };
